libs/termopol_db/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "termopol-db"
version = "0.1.0"
description = "Reusable database library for Termômetro da Polarização"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Frederico dos Santos Baker", email = "fredericodsbaker@gmail.com"}
]

dependencies = [
    "psycopg2-binary>=2.9.0",
    "pyyaml>=6.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[tool.setuptools]
packages = ["termopol_db"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
